requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0
chromadb>=0.4.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
